+++ tests/Endpoint/MyAccountsEndpointTest.php
@@ -11,17 +11,29 @@ class MyAccountsEndpointTest extends EndpointTestCase
 {
     public function testList(): void
     {
-        $this->endpoint()->list();
+        self::endpoint()->list();
         self::assertLastRequest('GET', '/api/my/accounts');
     }
 
     public function testCreate(): void
     {
-        $this->endpoint()->create(['foo' => 'bar']);
+        self::endpoint()->create(['foo' => 'bar']);
         self::assertLastRequest('POST', "/api/my/accounts", ['foo' => 'bar']);
     }
 
-    private function endpoint(): MyAccountsEndpoint
+    public function testAccept(): void
+    {
+        self::endpoint()->accept('foo');
+        self::assertLastRequest('POST', '/api/my/accounts/foo/accept');
+    }
+
+    public function testDecline(): void
+    {
+        self::endpoint()->decline('foo');
+        self::assertLastRequest('POST', '/api/my/accounts/foo/decline');
+    }
+
+    private static function endpoint(): MyAccountsEndpoint
     {
         return self::dgs()->my()->accounts();
     }